Change adjective fields
The is-adjective and adjective-type fields have been replaced with:

- is-adjective-i
- is-adjective-na

Additionally, there's the field:

- is-noun-and-na-adjective

This is because there are many words that can act both as a nound and
as a na-adjective. The above field causes a corresponding message to be
included in the card.

A word that can act both as a nound and a na-adjective must have all of
the is-adjective-na, is-noun, and is-noun-and-na-adjective fields set.